// src/state/types.ts
export type SerializedBigNumber = string

interface SerializedFarmUserData {
  allowance: string
  tokenBalance: string
  stakedBalance: string
  earnings: string
}

export interface DeserializedFarmUserData {
  allowance: BigNumber
  tokenBalance: BigNumber
  stakedBalance: BigNumber
  earnings: BigNumber
}

interface SerializedNFTFarmUserData {
  allowance: string
  nftBalance: string
  stakedBalance: string
  earnings: string
}
// ...
export interface SerializedNFTFarm extends SerializedNFTFarmConfig {
  earningTokenPriceBusd?: string
  nftAmountTotal?: SerializedBigNumber
  lpTotalInQuoteToken?: SerializedBigNumber
  lpTotalSupply?: SerializedBigNumber
  tokenPriceVsQuote?: SerializedBigNumber
  poolWeight?: SerializedBigNumber
  userData?: SerializedNFTFarmUserData
}
// ...
export interface SerializedNFTFarmsState {
  data: SerializedNFTFarm[]
  loadArchivedFarmsData: boolean
  userDataLoaded: boolean
  loadingKeys: Record<string, boolean>
  poolLength?: number
}